Bug 1959977 - Collect sourceId of JS interpreter and JS tracer frames r?iain,mstange
This patch collects the sourceIds of JS interpreter and JS execution
tracer frames inside the buffer. Then it serializes them as an index
into source table in frameTable location field, in a form like:
funcName (https://example.com/main.js:12:34)[<sourceIndex>]

This is made this way to make sure that we don't break the older Firefox
versions, so frontend will always treat `[sourceIndex]` as optional.

Differential Revision: https://phabricator.services.mozilla.com/D259254