Bug 1959977 - Collect sourceId of JIT frames r?iain,mstange
Thanks to the patches from Iain before, we now have ScriptSource inside
the JitCodeGlobalEntries. So we directly retrieve the sourceIds from
there during serialization (because we know that they are kept alive
until then).

And serialization into the profile format is done the same way as index
into sources table, by adding a `[sourceIndex]` at the end of the
location.

Differential Revision: https://phabricator.services.mozilla.com/D259255