Bug 1979114 - Propagate the profile generation information for child processes instead of duplicating them r?mstange
Previously, we were generating the SharedLibraries inside the
profiler_get_profile_json and then we were trying to generate it again
inside the caller instead of using the Result value from it. Now, this
patch changes this to use the additional information we generate inside
the caller instead of generating it multiple times.

That way:
1. We don't generate the same information multiple times.
2. The information that we generated in the profiler calls inside
   platform.cpp file is not going to be lost.

Differential Revision: https://phabricator.services.mozilla.com/D259262