Bug 1441689 - Add line and column information to the JIT frames r?iain!,mstange!
This patch adds the line and column numbers to the profiler JIT frames
by returning them in the callStackAtAddr methods.

Also it makes sure that the hash function contains the line and column
information, so we don't mistakenly merge multiple frames with different
line/column numbers together.

Another important change is the `wantPreciseLineNumbers` variable to make
sure that the Ion code can contain information for precise line numbers.

Note that with this patch, Ion will contain correct line numbers but not
Baseline. While working on this, we discovered that the baseline line
numbers are very sparse due to the fact that compiled Baseline assembly
doesn't contain profilerCallSite information. It will be fixed in the
following patches. And the test for this is also in one of the following
patches.

Co-authored-by: Markus Stange <mstange.moz@gmail.com>
Co-authored-by: Nicolas B. Pierron <nicolas.b.pierron@mozilla.com>
Co-authored-by: Denis Palmeiro <dpalmeiro@mozilla.com>

Differential Revision: https://phabricator.services.mozilla.com/D270205