Bug 1441689 - Add JSScript back to IonEntry/BaselineEntry r?iain
This patch does a partial backout of https://phabricator.services.mozilla.com/D255393
because we removed the JSScript at first thinking that we don't need it
while we are profiling. But we need it to be able to get the line and
column information for JIT frames. See the following patch to see how we
use the JSScript inside callStackAtAddr.

It looks like there is another way to get this line/column information
which is using utilizing the data that we keep for PerfSpewer. But
unfortunately that doesn't work well for us for now because PerfSpewer
doesn't know about the inlined JIT frames. If we add this information to
PerfSpewer, then we can potentially revert this patch again. But I would
like to work on this iteratively, by having this information first, and
then improving it later.

Differential Revision: https://phabricator.services.mozilla.com/D275593