Bug 1441689 - Keep a copy of CallStackFrameInfo directly in ProfiledFrameHandle r?iain
Previously we were keeping the fields of CallStackFrameInfo one by one
in the ProfiledFrameHandle. But since we added this struct now, we can
simply pass it and keep it around instead. This is going to be useful
when we add more values to this struct like the line and column number.

Differential Revision: https://phabricator.services.mozilla.com/D275594