Bug 1441689 - Add profiler jit-test for JS frame line numbers r?iain,jandem
Now that we have fixed the profiler JIT line number attribution
properly, let's add a test to verify this.

This patch simulates the failure case that I was experiencing with the
baseline frames in one of my demo websites. The failure case had a
for-of loop and had a 3-4 calls inside of it that calls different DOM
APIs etc. The line numbers for all of these calls inside the for loop
were attributed to the for loop header itself, assuming because we had
profiler call sites emitted for iterator .next, but not for the others.

This test was failing before my commit that fixes the baseline profiler
call sites.

Differential Revision: https://phabricator.services.mozilla.com/D275597